Trait diesel::expression::Expression

pub trait Expression {
    type SqlType;
}

Represents a typed fragment of SQL.

Apps should not need to implement this type directly, but it may be common to use this in where clauses. Libraries should consider using diesel_infix_operator! or diesel_postfix_operator! instead of implementing this directly.

Associated Types

type SqlType

The type that this expression represents in SQL